update test to provide input parameter
update the test to call the create_database method with an input parameter for the path of where the database will be created. CURRENT_PATH global variable created to facilitate that. DATABASE_PATH global variable renamed to TEST_DATABASE_PATH to reflect that the path is the location of the test database